Refactor 4337 structure
