+++ packages/wallet/core/src/relayer/index.ts
@@ -1,8 +1,7 @@
+// Export the core interfaces and type guards
 export * from './relayer.js'
 export * from './bundler.js'
-export * as Local from './local.js'
-export * as Pk from './pk-relayer.js'
-export * as Rpc from './rpc/index.js'
-export * as Sequence from './sequence.js'
-export * as Erc4337 from './pimlico.js'
-export * as EIP6963 from './eip6963.js'
+
+// Group and export implementations
+export * as Standard from './standard/index.js'
+export * as Bundlers from './bundlers/index.js'
