+++ packages/wallet/core/src/relayer/standard/sequence.ts
@@ -1,7 +1,7 @@
 import { ETHTxnStatus, IntentPrecondition, Relayer as Service } from '@0xsequence/relayer'
 import { Payload } from '@0xsequence/wallet-primitives'
 import { AbiFunction, Address, Bytes, Hex } from 'ox'
-import { FeeOption, FeeQuote, OperationStatus, Relayer } from './relayer.js'
+import { FeeOption, FeeQuote, OperationStatus, Relayer } from '../relayer.js'
 
 export class SequenceRelayer implements Relayer {
   public readonly kind: 'relayer' = 'relayer'
