+++ packages/wallet/wdk/src/sequence/types/index.ts
@@ -18,7 +18,7 @@ export type { Kind, RecoverySigner, SignerWithKind, WitnessExtraSignerKind } fro
 export type {
   BaseRelayerOption,
   ERC4337RelayerOption,
-  LegacyRelayerOption,
+  StandardRelayerOption,
   RelayerOption,
   Transaction,
   TransactionDefined,
